refactor: Adjust tests to latest changes
