+++ apps/api/src/paypal/paypal.controller.spec.ts
@@ -8,6 +8,8 @@ import { HttpModule } from '@nestjs/axios'
 import { NotificationModule } from '../sockets/notifications/notification.module'
 
 import { MarketingNotificationsModule } from '../notifications/notifications.module'
+import { DonationsModule } from '../donations/donations.module'
+import { CACHE_MANAGER, CacheModule } from '@nestjs/cache-manager'
 
 describe('PaypalController', () => {
   let controller: PaypalController
@@ -15,15 +17,17 @@ describe('PaypalController', () => {
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       imports: [
+        CacheModule.register({ isGlobal: true }),
         PaypalModule,
         CampaignModule,
         MarketingNotificationsModule,
         ConfigModule,
         HttpModule,
+        DonationsModule,
         NotificationModule,
       ],
       controllers: [PaypalController],
-      providers: [PaypalService],
+      providers: [PaypalService, { provide: CACHE_MANAGER, useValue: {} }],
     }).compile()
 
     controller = module.get<PaypalController>(PaypalController)
