+++ apps/api/src/stripe/events/stripe-payment.service.spec.ts
@@ -55,12 +55,18 @@ import { MarketingNotificationsService } from '../../notifications/notifications
 import { EmailService } from '../../email/email.service'
 import { TemplateService } from '../../email/template.service'
 import type { PaymentWithDonation } from '../../donations/types/donation'
+import { DonationsService } from '../../donations/donations.service'
+import { StripeService } from '../stripe.service'
+import { ExportService } from '../../export/export.service'
 
 const defaultStripeWebhookEndpoint = '/stripe/webhook'
 const stripeSecret = 'wh_123'
 
 describe('StripePaymentService', () => {
   let stripePaymentService: StripePaymentService
+  let campaignService: CampaignService
+  let donationService: DonationsService
+  let stripeService: StripeService
   let app: INestApplication
   const stripe = new Stripe(stripeSecret, { apiVersion: '2022-11-15' })
 
@@ -136,6 +142,9 @@ describe('StripePaymentService', () => {
         VaultService,
         PersonService,
         RecurringDonationService,
+        StripeService,
+        DonationsService,
+        ExportService,
         {
           provide: HttpService,
           useValue: mockDeep<HttpService>(),
@@ -150,6 +159,9 @@ describe('StripePaymentService', () => {
     await app.init()
 
     stripePaymentService = app.get<StripePaymentService>(StripePaymentService)
+    donationService = app.get<DonationsService>(DonationsService)
+    campaignService = app.get<CampaignService>(CampaignService)
+    stripeService = app.get<StripeService>(StripeService)
 
     //this intercepts the request raw body and removes the exact signature check
     const stripePayloadService = app.get<StripePayloadService>(StripePayloadService)
@@ -173,7 +185,6 @@ describe('StripePaymentService', () => {
       secret: stripeSecret,
     })
 
-    const campaignService = app.get<CampaignService>(CampaignService)
     const mockedCampaignById = jest
       .spyOn(campaignService, 'getCampaignById')
       .mockImplementation(() => Promise.resolve(mockedCampaign))
@@ -181,7 +192,7 @@ describe('StripePaymentService', () => {
     const paymentData = getPaymentData(mockPaymentEventCreated.data.object as Stripe.PaymentIntent)
 
     const mockedUpdateDonationPayment = jest
-      .spyOn(campaignService, 'updateDonationPayment')
+      .spyOn(donationService, 'updateDonationPayment')
       .mockImplementation(() => Promise.resolve(''))
       .mockName('updateDonationPayment')
 
@@ -214,7 +225,6 @@ describe('StripePaymentService', () => {
       secret: stripeSecret,
     })
 
-    const campaignService = app.get<CampaignService>(CampaignService)
     const mockedCampaignById = jest
       .spyOn(campaignService, 'getCampaignById')
       .mockImplementation(() => Promise.resolve(mockedCampaign))
@@ -224,7 +234,7 @@ describe('StripePaymentService', () => {
     )
 
     const mockedUpdateDonationPayment = jest
-      .spyOn(campaignService, 'updateDonationPayment')
+      .spyOn(donationService, 'updateDonationPayment')
       .mockImplementation(() => Promise.resolve(''))
       .mockName('updateDonationPayment')
 
@@ -252,7 +262,6 @@ describe('StripePaymentService', () => {
       secret: stripeSecret,
     })
 
-    const campaignService = app.get<CampaignService>(CampaignService)
     const mockedCampaignById = jest
       .spyOn(campaignService, 'getCampaignById')
       .mockImplementation(() => Promise.resolve(mockedCampaign))
@@ -260,7 +269,7 @@ describe('StripePaymentService', () => {
     const paymentData = getPaymentData(mockPaymentEventFailed.data.object as Stripe.PaymentIntent)
 
     const mockedUpdateDonationPayment = jest
-      .spyOn(campaignService, 'updateDonationPayment')
+      .spyOn(donationService, 'updateDonationPayment')
       .mockImplementation(() => Promise.resolve(''))
       .mockName('updateDonationPayment')
 
@@ -291,7 +300,6 @@ describe('StripePaymentService', () => {
       secret: stripeSecret,
     })
 
-    const campaignService = app.get<CampaignService>(CampaignService)
     const vaultService = app.get<VaultService>(VaultService)
 
     const mockedCampaignById = jest
@@ -326,7 +334,7 @@ describe('StripePaymentService', () => {
 
     jest.spyOn(prismaMock, '$transaction').mockImplementation((callback) => callback(prismaMock))
     const mockedUpdateDonationPayment = jest
-      .spyOn(campaignService, 'updateDonationPayment')
+      .spyOn(donationService, 'updateDonationPayment')
       .mockName('updateDonationPayment')
 
     const mockedIncrementVaultAmount = jest.spyOn(vaultService, 'incrementVaultAmount')
@@ -401,7 +409,7 @@ describe('StripePaymentService', () => {
 
     jest.spyOn(prismaMock, '$transaction').mockImplementation((callback) => callback(prismaMock))
     const mockedUpdateDonationPayment = jest
-      .spyOn(campaignService, 'updateDonationPayment')
+      .spyOn(donationService, 'updateDonationPayment')
       .mockName('updateDonationPayment')
 
     const mockedIncrementVaultAmount = jest.spyOn(vaultService, 'incrementVaultAmount')
@@ -461,7 +469,7 @@ describe('StripePaymentService', () => {
 
     jest.spyOn(prismaMock, '$transaction').mockImplementation((callback) => callback(prismaMock))
     const mockedUpdateDonationPayment = jest
-      .spyOn(campaignService, 'updateDonationPayment')
+      .spyOn(donationService, 'updateDonationPayment')
       .mockName('updateDonationPayment')
 
     const mockDecremementVaultAmount = jest.spyOn(vaultService, 'decrementVaultAmount')
@@ -574,7 +582,7 @@ describe('StripePaymentService', () => {
     jest.spyOn(prismaMock, '$transaction').mockImplementation((callback) => callback(prismaMock))
 
     const mockedUpdateDonationPayment = jest
-      .spyOn(campaignService, 'updateDonationPayment')
+      .spyOn(donationService, 'updateDonationPayment')
       .mockName('updateDonationPayment')
 
     prismaMock.payment.findFirst.mockResolvedValue({
@@ -621,7 +629,7 @@ describe('StripePaymentService', () => {
     const recurring = app.get<RecurringDonationService>(RecurringDonationService)
 
     const mockCancelSubscription = jest
-      .spyOn(recurring, 'cancelSubscription')
+      .spyOn(stripeService, 'cancelSubscription')
       .mockImplementation(() => Promise.resolve(null))
 
     const mockedCampaignById = jest
@@ -629,7 +637,7 @@ describe('StripePaymentService', () => {
       .mockImplementation(() => Promise.resolve(mockedCampaignCompeleted))
 
     const mockedUpdateDonationPayment = jest
-      .spyOn(campaignService, 'updateDonationPayment')
+      .spyOn(donationService, 'updateDonationPayment')
       .mockImplementation(() => Promise.resolve(''))
       .mockName('updateDonationPayment')
