+++ apps/api/src/stripe/stripe.module.ts
@@ -2,13 +2,16 @@ import { Module } from '@nestjs/common'
 import { StripeModule as StripeClientModule } from '@golevelup/nestjs-stripe'
 import { StripeService } from './stripe.service'
 import { StripeController } from './stripe.controller'
-import { ConfigService } from '@nestjs/config'
+import { ConfigModule, ConfigService } from '@nestjs/config'
 import { StripeConfigFactory } from '../donations/helpers/stripe-config-factory'
 import { CampaignModule } from '../campaign/campaign.module'
 import { PersonModule } from '../person/person.module'
 
 import { DonationsModule } from '../donations/donations.module'
 import { RecurringDonationModule } from '../recurring-donation/recurring-donation.module'
+import { StripePaymentService } from './events/stripe-payment.service'
+import { EmailService } from '../email/email.service'
+import { TemplateService } from '../email/template.service'
 
 @Module({
   imports: [
@@ -16,12 +19,13 @@ import { RecurringDonationModule } from '../recurring-donation/recurring-donatio
       inject: [ConfigService],
       useFactory: StripeConfigFactory.useFactory,
     }),
+    ConfigModule,
     CampaignModule,
     PersonModule,
     DonationsModule,
     RecurringDonationModule,
   ],
-  providers: [StripeService],
+  providers: [StripeService, StripePaymentService, EmailService, TemplateService],
   controllers: [StripeController],
 })
 export class StripeModule {}
