+++ package.json
@@ -26,7 +26,7 @@
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.369.0",
-    "@golevelup/nestjs-stripe": "0.6.0",
+    "@golevelup/nestjs-stripe": "^0.7.0",
     "@golevelup/nestjs-webhooks": "0.2.14",
     "@keycloak/keycloak-admin-client": "18.0.0",
     "@nestjs/axios": "^2.0.0",
